Check remove conditions on single line

+++ src/safm_api/views.py
@@ -361,10 +361,9 @@ class UserProfileView(generics.GenericAPIView):
                     serializer.is_valid(raise_exception=True)
     
                     # Removes the old profile picture if a new one is uploaded
-                    if request.data.get('profile_picture'):
-                        if not profile.has_default_picture:
-                            path_old_picture = os.path.join(settings.MEDIA_ROOT, profile.profile_picture.name)
-                            os.remove(path_old_picture)
+                    if request.data.get('profile_picture') and not profile.has_default_picture:
+                        path_old_picture = os.path.join(settings.MEDIA_ROOT, profile.profile_picture.name)
+                        os.remove(path_old_picture)
 
                     profile = serializer.save()
